I added a few macros and moved all the others.
I made a Bed Eject and Bed Clean Macro to eject the part and roll the bed for easier cleaning and moved all the gcode_macro files into this one file to make them easier to find and clean up the printer.cfg file.  You will want to delete the files out of the printer.cfg file and add the [include macros.cfg] to the top of your printer config or use the printer.cfg file I have in the main branch NOT the stock or demo cfg files.